happier envelopes
offload envelope slope handling into some of the spare cycles the output
cog has. This is a massive improvement for both the WTF'ness of the Spin
side of synth.env and the overall sound of the synth.

The duration values are a bit more sensible for the 00-99 range as well. We
can go almost 2.5 minutes with value 0 (The DX7 was about 3 minutes) and
the linearity is similar. Lack of L/R scaling aside, transcribing DX7
envelopes should be pretty close.